+++ src-api/app/main.py
@@ -8,32 +8,28 @@ from pydantic import ValidationError
 from cloudant import CouchDB
 from cloudant.error import CloudantDatabaseException
 
+from studium.core.datacontroller import Database
 from studium.core.models import User, Token, NewUser, UserInDB
-from studium.core.security import authenticate_user, get_current_active_user, is_valid_email, get_password_hash
+from studium.core.security import authenticate_user, get_current_active_user, get_password_hash
 from studium.core.jwt import create_access_token, ACCESS_TOKEN_EXPIRE_MINUTES
 
 sentry_dsn = os.getenv("SENTRY_DSN")
 sentry_sdk.init(sentry_dsn, traces_sample_rate=1.0)
 
-# conn = DataController()
-# conn.get_database('_all_dbs')
-# conn.create_partition_db('ok')
+
 client = CouchDB(**{
     'user': os.getenv("COUCHDB_USER"),
     'auth_token': os.getenv("COUCHDB_PASSWORD"),
     'url': os.getenv("COUCHDB_URL"),
     'connect': True})
 
-
 client.connect()
 
-try:
-    users_db = client['users']
-except Exception as e:
-    print(e)
-    users_db = 'users'
-    print(f"[CouchDB] Creating database: {users_db}")
-    users_db = client.create_database(users_db, partitioned=False)
+# Database connection
+db = Database()
+
+users_db = db.get_users()
+print(users_db['jerlends@tuta.io'].exists())
 
 
 app = FastAPI()
@@ -41,15 +37,9 @@ app = FastAPI()
 
 @app.post("/account/create")
 async def create_account(form_data: NewUser):
-    unknown_email = is_valid_email(form_data.email)
-    if not unknown_email['is_valid']:
-        return {'error': unknown_email['value']}
-
-    email = unknown_email['value']
 
     if form_data.password_one != form_data.password_two:
         return {'error': 'Passwords do not match.'}
-
     password = form_data.password_one
 
     if len(password) < 8:
@@ -60,13 +50,15 @@ async def create_account(form_data: NewUser):
     account_creation_time = str(datetime.now())
 
     new_user = {
-        '_id': email,
+        '_id': form_data.email,
         'username': form_data.username,
         'hashed_password': hashed_password,
         'ctime': account_creation_time,
     }
+
     try:
         user = UserInDB(**new_user)
+
     except ValidationError as e:
         # TODO: Implement logs
         print(e.json())
@@ -84,14 +76,17 @@ async def create_account(form_data: NewUser):
 
 @app.post("/token", response_model=Token)
 async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
-    user = authenticate_user({'username': 'admin', 'password': 'password'}, form_data.username, form_data.password)
+    # username == _id == email
+    user = authenticate_user(form_data.username, form_data.password)
+    print(user)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
+            detail="Incorrect email or password",
             headers={"WWW-Authenticate": "Bearer"},
         )
     access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    print(user.username)
     access_token = create_access_token(
         data={"sub": user.username}, expires_delta=access_token_expires
     )
@@ -103,7 +98,7 @@ async def read_users_me(current_user: User = Depends(get_current_active_user)):
     return current_user
 
 
-@app.get("/users/me/items/")
+@app.get("/users/account}/")
 async def read_own_items(current_user: User = Depends(get_current_active_user)):
     return [{"item_id": "Foo", "owner": current_user.username}]