+++ src-api/app/studium/core/security.py
@@ -2,37 +2,16 @@ from passlib.context import CryptContext
 from jose import JWTError, jwt
 from fastapi.security import OAuth2PasswordBearer
 from fastapi import Depends, HTTPException, status
-from email_validator import validate_email, EmailNotValidError
 
 from .jwt import SECRET_KEY, ALGORITHM
 from .models import UserInDB, TokenData, User
+from .datacontroller import Database
 
-fake_users_db = {}  # Todo: implement cloudant
 
+database = Database()
+users_db = database.get_users()
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-def is_valid_email(unknown_email: str) -> dict:
-    """https://pypi.org/project/email-validator/
-    Validates an email and returns dict with normalized form or human-readable error message.
-
-    dict keys: is_valid, value (email or error message)
-    :param unknown_email: an email address
-    :return:
-    """
-    try:
-        email_address = validate_email(unknown_email)
-        email = email_address.email
-        return {
-            'value': email,
-            'is_valid': True
-        }
-    except EmailNotValidError as e:
-        return {
-            'value': str(e),
-            'is_valid': False
-        }
 
 
 def verify_password(plain_password, hashed_password):
@@ -43,15 +22,18 @@ def get_password_hash(password):
     return pwd_context.hash(password)
 
 
-def get_user(db, username: str):
-    if username in db:
-        user_dict = db[username]
+def get_user(user_email: str):
+    try:
+        user_dict = users_db[user_email]
         return UserInDB(**user_dict)
+    except Exception as e:
+        print(e)
 
 
-def authenticate_user(database, username: str, password: str):
-    user = get_user(database, username)
+def authenticate_user(user_email: str, password: str):
+    user = get_user(user_email)
     if not user:
+        print(user)
         return False
     if not verify_password(password, user.hashed_password):
         return False
@@ -69,13 +51,13 @@ async def get_current_user(token: str = Depends(oauth2_scheme)):
     )
     try:
         payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
+        user_email: str = payload.get("sub")
+        if user_email is None:
             raise credentials_exception
-        token_data = TokenData(username=username)
+        token_data = TokenData(user_email=user_email)
     except JWTError:
         raise credentials_exception
-    user = get_user(fake_users_db, username=token_data.username)
+    user = get_user(user_email=token_data.user_email)
     if user is None:
         raise credentials_exception
     return user