+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Connected login endpoints to database
 - Added dependency email-validator
 - FastAPI account creation endpoint 2021-04-19
 - FastAPI dependency w/ intention of creating public api 2021-04-17
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Removed
+- Removed email-validator to use pydantics EmailStr
 - Database classes. These will be reimplemented once I determine common uses
 [Unreleased]: https://github.com/jerlendds/studium